+++ docs/changelog.md
@@ -54,6 +54,10 @@
 
 ## 更新说明
 
+1.9.30 fix: fix stock_a_lg_indicator interface
+
+    1. 修复 stock_a_lg_indicator 接口 
+
 1.9.29 fix: fix stock_a_lg_indicator interface
 
     1. 修复 stock_a_lg_indicator 接口
@@ -2079,6 +2083,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.30 fix: fix stock_a_lg_indicator interface
 
 1.9.29 fix: fix stock_a_lg_indicator interface